% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationMetricsTools.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy Function}
\usage{
accuracy(y_pred, y_true)
}
\arguments{
\item{y_pred}{array-like of shape (n_samples,) Estimated targets as returned by a classifier.}

\item{y_true}{array-like of shape (n_samples,) Ground truth (correct) target values.}
}
\value{
Outputs the Acccuracy
}
\description{
Calculates the accuracy  given the true and predicted values
}
\details{
Accuracy is defined as (TP +  TN) / (All Positives + All Negatives).
}
\examples{
Simulated y_true and y_predicted values
y_true = c(1,1,1,0,0,1,0,1,1,0)
y_pred = c(1,1,1,1,0,0,1,0,1,1)
accuracy(y_pred, y_true)

}
