% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted R Squared Function}
\usage{
adjR2(y_true, y_pred)
}
\arguments{
\item{y_true}{array-like of shape (n_samples,) Ground truth (correct) target values.}

\item{y_pred}{array-like of shape (n_samples,) Estimated target values}
}
\value{
outputs the r-square value
}
\description{
Calculates the adjusted r-squared value
}
\examples{
y_true = c(4, -1.5, 5, 9)
y_pred = c(3.5, 1.0, 5, 10)
adjR2(y_pred, y_true)

}
