% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{bhAdjust}
\alias{bhAdjust}
\title{Bonferroni Holm Function}
\usage{
bhAdjust(p_values, alpha = 0.5, method = "holm")
}
\arguments{
\item{p_values}{a vector of p_value}

\item{alpha}{alpha value to use for adjustment}

\item{method}{method for adjustment. Default is holm}
}
\description{
Performs an adjusted bonferroni correction
}
