% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{boxMuller}
\alias{boxMuller}
\title{BoxMuller Function}
\usage{
boxMuller(n = 500)
}
\arguments{
\item{n}{the number of samples to create normal distribution}
}
\value{
a normal distribution
}
\description{
The Box-Muller transformation takes two samples from the uniform distribution
on the interval \link{0, 1} and maps them to two standard, normally distributed samples:
рќ‘‹1=в€’2log(рќ‘€1)вЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕв€љcos(2рќњ‹рќ‘€2)
рќ‘‹2=в€’2log(рќ‘€1)вЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕвЂѕв€љsin(2рќњ‹рќ‘€2)
}
\examples{
n = 500
BM_data = boxMuller(n)
find means of boxMuller
round(mean(BM_data[,1]),2)
find variance of boxMuller  
round(var(BM_data[,1]),2) 

}
