% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{chiSquareCont}
\alias{chiSquareCont}
\title{Chi Square Test}
\usage{
chiSquareCont(tib)
}
\arguments{
\item{tib}{data supplied as a contingency table}
}
\value{
It returns the test statistics, the p-value and the degree of freedom
}
\description{
calculates a chi-square test for count data supplied as a contingency table
}
