% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{fdrAdjust}
\alias{fdrAdjust}
\title{FDR Function}
\usage{
fdrAdjust(p_values, alpha = 0.5, method = "fdr")
}
\arguments{
\item{p_values}{a vector of p_value}

\item{alpha}{alpha value to use for adjustment}

\item{method}{method for adjustment. Default is FDR}
}
\description{
Performs an adjusted FDR correction
}
\examples{
Lets say we have performed a series of statistical tests 
and obtained the following p-values:
p.vals = seq(0.0025,0.025,0.0025)

We could give them names, to identify the results of each test.
names(p.vals) = paste("Test",LETTERS[1:10])

fdrAdjust(p_values = p.vals, alpha = 0.5, method = "fdr")

}
