% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{minimumN}
\alias{minimumN}
\title{Minimum Sample Size Function}
\usage{
minimumN(d = 0.8)
}
\arguments{
\item{d}{the effect size}
}
\value{
the minimum sample size
}
\description{
This function finds the minimum (per group) sample size when given the effect size
}
\examples{
minimumN(0.6)

}
