% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaHelperTools.R
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Approximation}
\usage{
pcApprox(X, npc)
}
\arguments{
\item{X}{tibble or matrix}

\item{npc}{the number of pca components}
}
\value{
An approximated tibble of the original data
}
\description{
This function returns an approximation to the data x based on the number of Principal
Components (PCs).
}
\details{
The closer the number of principal components to the number of features in the original data,
the better approximated the input value.
}
\examples{
npc = 6
pcApprox(mtcars[,1:npc],npc)

}
