% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaHelperTools.R
\name{pcLollipop}
\alias{pcLollipop}
\title{Lollipop Plot}
\usage{
pcLollipop(X)
}
\arguments{
\item{X}{tibble or matrix}
}
\value{
A lollipop plot of the principal component loadings
}
\description{
This function creates a lollipop plot of the principal
component loadings of the (potentially unscaled/uncentered) data x
}
\details{
The function scales the data before performing the principal component analysis.
}
\examples{
pcLollipop(mtcars[,1:6])

}
