% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{postHocPower}
\alias{postHocPower}
\title{Post Hoc Power Analysis}
\usage{
postHocPower(effect_size = 1, n1 = 30, n2 = 30)
}
\arguments{
\item{effect_size}{the effect size}

\item{n1, n2}{the number of samples per group}
}
\value{
outputs the post hoc power
}
\description{
This function simulates post-hoc power
}
\examples{
N = 1000
data_power = rep(NA,N)
for (i in 1:N) {data_power[i] = postHocPower(i)}

or
res = replicate(100,postHocPower(effect_size=1, n1=30, n2=30))
mean(res)

}
