% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationMetricsTools.R
\name{ppv}
\alias{ppv}
\title{Positive Predictive Value or Precision Function}
\usage{
ppv(y_pred, y_true)
}
\arguments{
\item{y_pred}{array-like of shape (n_samples,) Estimated targets as returned by a classifier.}

\item{y_true}{array-like of shape (n_samples,) Ground truth (correct) target values.}
}
\value{
Outputs the precision
}
\description{
Calculates the Positive Predictive Value or precision given the true and predicted values
}
\details{
Positive Predictive Value, or PPV, is defined as True Positives / ( True Positives + False Positives).
PPV is also referred to as precision.
}
\examples{
Simulated y_true and y_predicted values
y_true = c(1,1,1,0,0,1,0,1,1,0)
y_pred = c(1,1,1,1,0,0,1,0,1,1)
ppv(y_pred, y_true)

}
