% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationMetricsTools.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity Function}
\usage{
sensitivity(y_pred, y_true)
}
\arguments{
\item{y_pred}{array-like of shape (n_samples,) Estimated targets as returned by a classifier.}

\item{y_true}{array-like of shape (n_samples,) Ground truth (correct) target values.}
}
\value{
Outputs the sensitivity
}
\description{
Compute the sensitivity given the true and predicted values
}
\details{
Sensitivity is defined as
True Positives / All Positives = True Positives / ( True Positives + False Negatives).
Sensitivity is also referred to as recall.
}
\examples{
Simulated y_true and y_predicted values
y_true = c(1,1,1,0,0,1,0,1,1,0)
y_pred = c(1,1,1,1,0,0,1,0,1,1)
sensitivity(y_pred,y_true)

}
