% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{simulateTrial}
\alias{simulateTrial}
\title{Simulate data}
\usage{
simulateTrial(N = 50)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a simulated with treatment and control groups
}
\description{
This function is used to generate data for the effect size function.
It creates a simulated data with treatment and control groups
}
\examples{
simulated_cancer = simulateTrial(50)

}
