% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{two_sided_z}
\alias{two_sided_z}
\title{Two Sided Z function}
\usage{
two_sided_z(z)
}
\arguments{
\item{z}{vector of quantiles.}
}
\description{
This function converts a test statistic x into the area under the z-density for values >|x|
Refer to the \code{pnorm} function for more details
}
\examples{
twoSidedT(2)

}
