% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{twoSidedT}
\alias{twoSidedT}
\title{Two Sided T function}
\usage{
twoSidedT(t, df)
}
\arguments{
\item{t}{vector of quantiles.}

\item{df}{degrees of freedom (> 0, maybe non-integer). df = Inf is allowed.}
}
\description{
This function converts a test statistic x into the area under the t-density for values >|x|
}
\details{
Refer to the \code{pt} function for more details
}
\examples{
twoSidedT(t=5, df=2)

}
