% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaHelperTools.R
\name{unscale}
\alias{unscale}
\title{Unscale Matrix-like Objects}
\usage{
unscale(X)
}
\arguments{
\item{X}{vector or matrix}
}
\value{
An unscaled vector or matrix
}
\description{
This function takes a numerical object from scale and reverses the centering/scaling.
It assumes the input is the output of the scale function in R
}
\examples{
 
x = c(2,4,3,5)
X = matrix(x, nrow = 2)
X_scaled = scale(X)
unscale(X_scaled)

}
