% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{welchT}
\alias{welchT}
\title{Welch T-test function}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{vector of values}

\item{y}{vector of values. (Should be the same length as x)}
}
\value{
It returns the test statistics, the p-value and the degree of freedom
}
\description{
Perform a welch t-test
}
\examples{
x = rnorm(10,0,1)
y = rnorm(10,1,1)
welchT(x,y)

}
