% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{Chi Square Test}
\usage{
chiSquareCounts(tib)
}
\arguments{
\item{tib}{data supplied as a contingency table}
}
\value{
outputs the chi square value
}
\description{
calculates a chi-square test for count data supplied as a contingency table
}
