% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestingTools.R
\name{effectSize}
\alias{effectSize}
\title{Effect Size Function}
\usage{
effectSize(t, g)
}
\arguments{
\item{t}{\itemize{
\item vector of data points
}}

\item{g}{\itemize{
\item groups for which each value in parameter t belongs
}}
}
\value{
outputs the Cohen’s d effect size
}
\description{
A function to calculate Cohen’s d effect size.
}
\details{
This is the difference in means between two groups
divided by some measure of the standard deviation.
}
\examples{
simulated_cancer = simulateTrial(50)
effectSize(simulated_cancer, simulated_cancer[,1]) 

}
