% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationMetricsTools.R
\name{F1}
\alias{F1}
\title{F1 Score Function}
\usage{
F1(y_pred, y_true)
}
\arguments{
\item{y_pred}{array-like of shape (n_samples,) Estimated targets as returned by a classifier.}

\item{y_true}{array-like of shape (n_samples,) Ground truth (correct) target values.}
}
\value{
Outputs the F1-score
}
\description{
Calculates the F1-Score given the true and predicted values
}
\details{
F1 score is defined as 2 x (precision x recall) / (precision + recall)
}
\examples{
Simulated y_true and y_predicted values
y_true = c(1,1,1,0,0,1,0,1,1,0)
y_pred = c(1,1,1,1,0,0,1,0,1,1)
F1(y_pred, y_true)

}
